+++ global/src/string.cpp
@@ -31,7 +31,7 @@ public:
         XAML_CATCH_RETURN()
     }
 
-    xaml_result XAML_CALL get_length(size_t* psize) noexcept override
+    xaml_result XAML_CALL get_length(int32_t* psize) noexcept override
     {
         *psize = m_str.size();
         return XAML_S_OK;
@@ -48,7 +48,7 @@ public:
 
     xaml_result XAML_CALL equals(xaml_string* str, bool* pres) noexcept override
     {
-        size_t rhs_length;
+        int32_t rhs_length;
         xaml_result hr = str->get_length(&rhs_length);
         if (XAML_FAILED(hr)) return hr;
         xaml_char_t const* rhs_data;
